# proto/openapi/auv-daemon.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "auv/api/daemon/v1/resource.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/apis/auv/runtime/v1/runners": {
      "get": {
        "operationId": "RunnerService_ListRunners",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListRunnersResponse"
            }
          }
        },
        "tags": [
          "RunnerService"
        ]
      },
      "post": {
        "operationId": "RunnerService_CreateRunner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateRunnerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateRunnerRequest"
            }
          }
        ],
        "tags": [
          "RunnerService"
        ]
      }
    },
    "/apis/auv/runtime/v1/runners:delete": {
      "post": {
        "operationId": "RunnerService_DeleteRunner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteRunnerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1DeleteRunnerRequest"
            }
          }
        ],
        "tags": [
          "RunnerService"
        ]
      }
    },
    "/apis/auv/runtime/v1/runners:get": {
      "post": {
        "summary": "NOTICE(tonic-rest-path): typed RunnerRef path projection is deferred until\nthe generator supports arbitrary nested resource-reference messages.",
        "operationId": "RunnerService_GetRunner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetRunnerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetRunnerRequest"
            }
          }
        ],
        "tags": [
          "RunnerService"
        ]
      }
    },
    "/apis/auv/runtime/v1/runs": {
      "get": {
        "operationId": "RunService_ListRuns",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListRunsResponse"
            }
          }
        },
        "tags": [
          "RunService"
        ]
      },
      "post": {
        "operationId": "RunService_CreateRun",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateRunResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateRunRequest"
            }
          }
        ],
        "tags": [
          "RunService"
        ]
      }
    },
    "/apis/auv/runtime/v1/runs:get": {
      "post": {
        "summary": "NOTICE(tonic-rest-path): typed RunRef path projection is deferred until\nthe generator supports arbitrary nested resource-reference messages.",
        "operationId": "RunService_GetRun",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetRunResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetRunRequest"
            }
          }
        ],
        "tags": [
          "RunService"
        ]
      }
    },
    "/apis/auv/runtime/v1/runs:stop": {
      "post": {
        "operationId": "RunService_StopRun",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1StopRunResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1StopRunRequest"
            }
          }
        ],
        "tags": [
          "RunService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "v1CreateRunRequest": {
      "type": "object",
      "properties": {
        "devices": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1DeviceRef"
          },
          "description": "Empty selects the implicit local Device."
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "v1CreateRunResponse": {
      "type": "object",
      "properties": {
        "run": {
          "$ref": "#/definitions/v1Run"
        }
      }
    },
    "v1CreateRunnerRequest": {
      "type": "object",
      "properties": {
        "device": {
          "$ref": "#/definitions/v1DeviceRef",
          "description": "Absence selects the implicit local Device."
        },
        "runnerClass": {
          "$ref": "#/definitions/v1RunnerClassRef"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "lifecycle": {
          "$ref": "#/definitions/v1RunnerLifecycle"
        },
        "idleTimeout": {
          "type": "string",
          "description": "Required only for RUNNER_LIFECYCLE_UNLESS_IDLE."
        }
      }
    },
    "v1CreateRunnerResponse": {
      "type": "object",
      "properties": {
        "runner": {
          "$ref": "#/definitions/v1Runner"
        }
      }
    },
    "v1DeleteRunnerRequest": {
      "type": "object",
      "properties": {
        "runner": {
          "$ref": "#/definitions/v1RunnerRef"
        },
        "gracePeriod": {
          "type": "string",
          "description": "When absent, graceful draining waits without a deadline. A present zero\nduration performs an immediate deadline check."
        },
        "force": {
          "type": "boolean",
          "description": "Immediately terminates a daemon-owned executable. Remote gRPC endpoints\nare never terminated by deleting their local Runner resource."
        }
      }
    },
    "v1DeleteRunnerResponse": {
      "type": "object",
      "properties": {
        "runner": {
          "$ref": "#/definitions/v1Runner"
        }
      }
    },
...
    "v1DeviceRef": {
      "type": "object",
      "properties": {
        "deviceId": {
          "type": "string"
        }
      }
    },
...
    "v1GetRunRequest": {
      "type": "object",
      "properties": {
        "run": {
          "$ref": "#/definitions/v1RunRef"
        }
      }
    },
    "v1GetRunResponse": {
      "type": "object",
      "properties": {
        "run": {
          "$ref": "#/definitions/v1Run"
        }
      }
    },
...
    "v1GetRunnerRequest": {
      "type": "object",
      "properties": {
        "runner": {
          "$ref": "#/definitions/v1RunnerRef"
        }
      }
    },
    "v1GetRunnerResponse": {
      "type": "object",
      "properties": {
        "runner": {
          "$ref": "#/definitions/v1Runner"
        }
      }
    },
...
    "v1ListRunnersResponse": {
      "type": "object",
      "properties": {
        "runners": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Runner"
          }
        }
      }
    },
    "v1ListRunsResponse": {
      "type": "object",
      "properties": {
        "runs": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Run"
          }
        }
      }
    },
...
    "v1Run": {
      "type": "object",
      "properties": {
        "ref": {
          "$ref": "#/definitions/v1RunRef"
        },
        "phase": {
          "$ref": "#/definitions/v1RunPhase"
        },
        "devices": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1DeviceRef"
          }
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "completedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1RunOutcome": {
      "type": "string",
      "enum": [
        "RUN_OUTCOME_UNSPECIFIED",
        "RUN_OUTCOME_SUCCEEDED",
        "RUN_OUTCOME_FAILED",
        "RUN_OUTCOME_CANCELED"
      ],
      "default": "RUN_OUTCOME_UNSPECIFIED"
    },
    "v1RunPhase": {
      "type": "string",
      "enum": [
        "RUN_PHASE_UNSPECIFIED",
        "RUN_PHASE_PENDING",
        "RUN_PHASE_RUNNING",
        "RUN_PHASE_SUCCEEDED",
        "RUN_PHASE_FAILED",
        "RUN_PHASE_CANCELED"
      ],
      "default": "RUN_PHASE_UNSPECIFIED"
    },
    "v1RunRef": {
      "type": "object",
      "properties": {
        "runId": {
          "type": "string"
        }
      }
    },
    "v1Runner": {
      "type": "object",
      "properties": {
        "ref": {
          "$ref": "#/definitions/v1RunnerRef"
        },
        "device": {
          "$ref": "#/definitions/v1DeviceRef"
        },
        "runnerClass": {
          "$ref": "#/definitions/v1RunnerClassRef"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "lifecycle": {
          "$ref": "#/definitions/v1RunnerLifecycle"
        },
        "idleTimeout": {
          "type": "string"
        },
        "phase": {
          "$ref": "#/definitions/v1RunnerPhase"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "processId": {
          "type": "integer",
          "format": "int64",
          "description": "OS process identifier for daemon-owned process Runners. This is\ndiagnostic evidence, not a stable resource identity."
        },
        "activeOperations": {
          "type": "string",
          "format": "uint64"
        },
        "idleDeadline": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "v1RunnerClassRef": {
      "type": "object",
      "properties": {
        "runnerClass": {
          "type": "string"
        }
      }
    },
    "v1RunnerLifecycle": {
      "type": "string",
      "enum": [
        "RUNNER_LIFECYCLE_UNSPECIFIED",
        "RUNNER_LIFECYCLE_EPHEMERAL",
        "RUNNER_LIFECYCLE_UNLESS_IDLE",
        "RUNNER_LIFECYCLE_UNLESS_SHUTDOWN"
      ],
      "default": "RUNNER_LIFECYCLE_UNSPECIFIED"
    },
    "v1RunnerPhase": {
      "type": "string",
      "enum": [
        "RUNNER_PHASE_UNSPECIFIED",
        "RUNNER_PHASE_STARTING",
        "RUNNER_PHASE_READY",
        "RUNNER_PHASE_DRAINING",
        "RUNNER_PHASE_STOPPED",
        "RUNNER_PHASE_FAILED"
      ],
      "default": "RUNNER_PHASE_UNSPECIFIED"
    },
    "v1RunnerRef": {
      "type": "object",
      "properties": {
        "runnerId": {
          "type": "string"
        }
      }
    },
...
    "v1StopRunRequest": {
      "type": "object",
      "properties": {
        "run": {
          "$ref": "#/definitions/v1RunRef"
        },
        "outcome": {
          "$ref": "#/definitions/v1RunOutcome"
        }
      }
    },
    "v1StopRunResponse": {
      "type": "object",
      "properties": {
        "run": {
          "$ref": "#/definitions/v1Run"
        }
      }
    },
...
  }
}